# dist/api.yml
...
openapi: 3.0.0
info:
  title: Simple Inventory API
  description: This is a simple API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
paths:
...
  /games/GameDocumentation/{gameID}:
    get:
      tags:
        - Games
      summary: Return the decumentation of the game
      parameters:
        - name: gameID
          in: path
          description: game id to return details about
          required: true
          style: simple
          explode: false
          schema:
            type: string
            example: 1
      responses:
        "200":
          description: Game documentation has been returned to client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/gameDocumentation"
        "400":
          description: The request to get the game documantation has been denied
  /games/LeagueManagment/addGame:
    post:
      tags:
        - Games
      summary: Add game to the league managment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                  example: 2021-8-23
                  # format: date
                hour:
                  type: string
                  example: "20:30:00"
                  # format: time
                away_team_id:
                  type: number
                  example: 939
                home_team_id:
                  type: number
                  example: 1020
                field:
                  type: string
                  example: Tedi
                referee_name:
                  type: string
                  example: Yossi Cohen
      responses:
        "200":
          description: The game added to the league

        "400":
          description: The requset to add game to league has been denied
  /games/LeagueManagment/addScore:
    post:
      tags:
        - Games
      summary: Add score to game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                game_id:
                  type: number
                  example: 2
                home_team_goal:
                  type: number
                  example: 2
                away_team_goal:
                  type: number
                  example: 1
      responses:
        "200":
          description: The game has been update with the score

        "400":
          description: The requset to add score to game has been denied
  /games/LeagueManagment/addEvent:
    post:
      tags:
        - Games
      summary: Add event schedule to game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                game_id:
                  type: number
                  example: 1
                date:
                  type: string
                  example: 2020-2-18
                hour:
                  type: string
                  example: "21:45:00"
                game_minute:
                  type: number
                  example: 37
                event_type:
                  type: string
                  example: red card
                player_id:
                  type: number
                  example: 82808

      responses:
        "200":
          description: Event has been added to the game

        "400":
          description: The requset to add event  to game has been denied
...
components:
  schemas:
...
    gameDocumentation:
      type: object
      properties:
        game_id:
          type: string
          example: "2"
        game_date:
          type: string
          example: "Tue Jul 21 2020"
        game_hour:
          type: string
          example: 20:30:00
        home_team:
          type: string
          example: Real Madrid
        away_team:
          type: string
          example: Barcellona
        home_team_id:
          type: number
          example: 1020
        away_team_id:
          type: number
          example: 85
        home_team_goal:
          type: number
          description: If null- means the game not occur yet
          example: 2
        away_team_goal:
          type: number
          description: If null- means the game not occur yet
          example: 1
        field:
          type: string
          example: "Lyngby Stadion"
        eventSchedule:
          type: object
          description: If null- means the game not occur yet. cointains the events in the game
...